requirements.txt
flask>=3.0.0
openai>=1.14.0
google-generativeai>=0.5.0
python-dotenv>=1.0.1
colorama>=0.4.6
